refactor: drop only conversations in teams
